python time

+++ framestore.sh
@@ -8,14 +8,19 @@
 service_base=/tmp/service_framestore
 service_file=$service_base/framestore.py
 lockfile=/var/lock/subsys/framestore
+
+BASE=https://badabing.firebaseio-demo.com
+INSTANCE_ID=`curl -s http://169.254.169.254/latest/meta-data/instance-id`
+URL=$BASE/framestores/$INSTANCE_ID.json
+
 alias exec="/usr/bin/python"
 
 case $1 in
 	start)
 		[ -d $service_base ] && rm -R $service_base
-        /usr/bin/git clone https://github.com/adrianloh/framestore.git $service_base
+        /usr/bin/git clone https://github.com/adrianloh/framestore.git $service_base 2>/dev/null 1>/dev/null
         touch $lockfile
-		exec $service_file &
+		nohup exec $service_file > /tmp/framstore.log
 		;;
 	restart)
 		kill -9 `ps ax | grep $service_file | grep -v grep | awk '{print $1}'`
@@ -30,9 +35,15 @@ case $1 in
 		fi
 		;;
     stop)
-		kill -9 `ps ax | grep $service_file | grep -v grep | awk '{print $1}'`
-		rm -R $service_base
-		rm -f $lockfile
-		echo Framestore is stopped
+		proc=`ps ax | grep $service_file | grep -v grep | awk '{print $1}'`
+		if [ -n "$proc" ]; then
+			kill -9 $proc
+			[ -d $service_base ] && rm -R $service_base
+			[ -f $lockfile ] && rm -f $lockfile
+			curl -sX DELETE $URL > /dev/null
+			echo Framestore is stopped
+		else
+			echo Framestore is already stopped
+		fi
 		;;
 esac
